added demo stuff

+++ controllers/demo.php
@@ -17,8 +17,8 @@ class demo extends Controller {
     }
 
     function randomNumber($from, $to) {
-        $from = (int)trim(htmlspecialchars($from));
-        $to = (int)trim(htmlspecialchars($to));
+        $from = trim(htmlspecialchars($from));
+        $to = trim(htmlspecialchars($to));
 
         if (!is_numeric($from) || !is_numeric($to)) {
             $this->jsonOutputError("wrong interval, check your inputs");
